+++ src/direct.py
@@ -1,79 +1,56 @@
-from pulp import *
+import time
+from docplex.mp.model import Model
+import pandas as pd
+from collections import namedtuple
 
 
-def direct_method(prob_name, is_max_query, dataframe, objective_var, dot_product_constraints, obj_constraint):
-    n, m = dataframe.shape
-    if is_max_query:
-        prob = LpProblem(prob_name, LpMaximize)
+def direct(df, minmax, **kwargs):
+    # Inputs: dataframe df, "min or "max" for objective, A_0='attribute name' for objective function,
+    #        c#=('attribute_name', L_k, U_k) which give bounds for each attribute with a constraint
+    #        (# can be 1,2,...,14 order doesn't matter), count_range=(L_c, U,c) which is bounds for count constraint
+    # Output: [solvable as boolean, df of selected tuples, objective value, runtime]
+    start = time.time()
+    solvable = False
+    solution = None
+    objective = None
+    A_0 = kwargs.pop('A_0', None)
+    count_constraint = kwargs.pop('count_constraint', None)
+    constraints = [k for k in kwargs.values()]
+    mdl = Model()
+    # Create binary variable for each tuple, named after its index in df
+    variables = []
+    for i in range(len(df)):
+        variables.append(mdl.binary_var("t_{}".format(i)))
+    # Add constraints
+    for ct in constraints:
+        name, L_k, U_k = ct
+        if L_k is not None:
+            mdl.add_constraint(mdl.dot(variables, df[name]) >= L_k)
+        if U_k is not None:
+            mdl.add_constraint(mdl.dot(variables, df[name]) <= U_k)
+    # Add count constraint
+    if count_constraint is not None:
+        L_c, U_c = count_constraint
+        if L_c is not None:
+            mdl.add_constraint(mdl.sum(variables) >= L_c)
+        if U_c is not None:
+            mdl.add_constraint(mdl.sum(variables) <= U_c)
+    # Set objective.  If A_0 is none, we maximize or minimize count(*)
+    if A_0 is not None:
+        mdl.set_objective(minmax, mdl.dot(variables, df[A_0]))
     else:
-        prob = LpProblem(prob_name, LpMinimize)
-    lp_variables = [LpVariable("t_{}".format(i + 1), 0, 1, cat='Integer') for i in range(n)]
-    prob += lpDot(lp_variables, dataframe[objective_var])
-    for var, constraint, val in dot_product_constraints:
-        if constraint == 'less_than':
-            prob += (lpDot(lp_variables, dataframe[var]) < val)
-        elif constraint == 'less_than_equal_to':
-            prob += (lpDot(lp_variables, dataframe[var]) <= val)
-        elif constraint == 'equal':
-            prob += (lpDot(lp_variables, dataframe[var]) == val)
-        elif constraint == 'greater_than':
-            prob += (lpDot(lp_variables, dataframe[var]) > val)
-        elif constraint == 'greater_than_equal_to':
-            prob += (lpDot(lp_variables, dataframe[var]) >= val)
-        else:
-            print("****************not correct constraint***************")
-            exit(1)
-    for constraint, val in obj_constraint:
-        if constraint == 'less_than':
-            prob += (lpSum(lp_variables) < val)
-        elif constraint == 'less_than_equal_to':
-            prob += (lpSum(lp_variables) <= val)
-        elif constraint == 'equal':
-            prob += (lpSum(lp_variables) == val)
-        elif constraint == 'greater_than':
-            prob += (lpSum(lp_variables) > val)
-        elif constraint == 'greater_than_equal_to':
-            prob += (lpSum(lp_variables) >= val)
-        else:
-            print("****************not correct constraint***************")
-            exit(1)
-
-    status = prob.solve()
-    print("status of solved problem {}".format(status))
-    if status == 1:
-        return [value(v) for i, v in enumerate(lp_variables) if v == 1]
-    return []
-
-
-def first_query(data):
-    prob_name = "first_query"
-    dot_product_constraints = [('sum_base_price', 'less_than_equal_to', 15469853.7043),
-                               ('sum_disc_price', 'less_than_equal_to', 45279795.0584),
-                               ('sum_charge', 'less_than_equal_to', 95250227.7918),
-                               ('avg_qty', 'less_than_equal_to', 50.353948653),
-                               ('avg_price', 'less_than_equal_to', 68677.5852459),
-                               ('avg_disc', 'less_than_equal_to', 0.110243522496),
-                               ('sum_qty', 'less_than_equal_to', 77782.028739)]
-    obj_constraint = [('greater_than_equal_to', 1)]
-    obj = 'count_order'
-    solution = direct_method(prob_name=prob_name,
-                             is_max_query=True,
-                             dataframe=data,
-                             objective_var=obj,
-                             dot_product_constraints=dot_product_constraints,
-                             obj_constraint=obj_constraint)
-    return solution
-
-
-def second_query(data):
-    prob_name = "second_query"
-    dot_product_constraints = [('p_size', 'less_than_equal_to', 8)]
-    obj = 'ps_min_supplycost'
-    obj_constraint = [('greater_than_equal_to', 1)]
-    solution = direct_method(prob_name=prob_name,
-                             is_max_query=True,
-                             dataframe=data,
-                             objective_var=obj,
-                             dot_product_constraints=dot_product_constraints,
-                             obj_constraint=obj_constraint)
-    return solution
+        mdl.set_objective(minmax, mdl.sum(variables))
+    # Solve model
+    solve = mdl.solve()
+    if solve is not None:
+        indices = []
+        solvable = True
+        for var in variables:
+            if var.solution_value == 1:
+                indices.append(int(var.name[2:]))
+        solution = df.iloc[indices]
+        objective = mdl.objective_value
+    end = time.time()
+    run_time = end - start
+    final = namedtuple("final", ["solvable", "solution", "objective", "run_time"])
+    return final(solvable, solution, objective, run_time)
